pyproject.toml
[tool.poetry]
name = "cuma-be"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "3.1.0"
elasticsearch = "7.11.0"
elasticsearch-dsl = "7.4.0"
python-dotenv = "1.0.1"
flask-cors = "5.0.0"
flask-restx = "1.3.0"
flask-swagger-ui = "4.11.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
